build: use TypeScript for next.config
